skywalking-demo/python-demo/requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
requests==2.26.0
Apache-skywalking==1.0.1